Fixed issues & added new features
Fixed validators that made it impossible to create or edit a movie
Changed colour of letters to make them readeable.
Fixed Shrek issue.
Now when an user needs to be logged in it redirects to login screen, fixing error 500 issue on favorites/
Added Api functionality with optimization issues that i dont know how to fix, because image storing is costly.
Added Tags Crud